consolidate volume functions into docker.go
Move ListVolumes function from internal/docker/volume.go to internal/docker/docker.go to consolidate Docker client methods in a single file. This improves code organization by keeping all Docker client functionality together rather than spreading it across multiple files.
